requirements.txt
Flask==2.1.1
Werkzeug==2.1.1
python-dotenv==1.0.0
requests==2.28.1
opencv-python==4.7.0.72
numpy==1.23.1
pydantic==2.4.0
langchain-openai==0.1.19
langchain-core>=0.2.24,<0.3.0
